Fix default test script to suit IDA::Daemon.
Read config with two options:
* proto:     http | https    (default "http" to allow APP get and basic.t to work)
* auth_mode: server | mutual (default "server")

I will add an extra check later for "mode" (option for hypnotoad, etc.) that
checks/enforces "https" and "mutual" settings if the mode is "production".

These other options are to make testing less of a chore. If the proto is "http",
I ignore the auth_mode setting, and I can:

* run a daemon at a given port and interact insecurely with it via
  a web browser (or APP get)
* run a daemon via a unix domain socket, for secure testing